requirements.txt
Flask==0.12.2
requests==2.18.4
Fabric3==1.13.1.post1
gunicorn==19.7.1
validators==0.12.0
redis==2.10.6
gevent==1.2.2

# test dependencies
behave==1.2.5
nose==1.3.7